+++ crates/lmdb-js-lite/__benchmarks__/write-throughput-safe.ts
@@ -32,10 +32,12 @@ async function main() {
     });
 
     {
+      // eslint-disable-next-line no-console
       console.log('Generating entries for testing');
       const entries = [...Array(NUM_ENTRIES)].map(() => {
         return generateEntry();
       });
+      // eslint-disable-next-line no-console
       console.log('(no-batching) Writing entries for', MAX_TIME, 'ms');
       const start = Date.now();
       let numEntriesInserted = 0;
@@ -49,6 +51,7 @@ async function main() {
       await safeDB.commitWriteTransaction();
       const duration = Date.now() - start;
       const throughput = numEntriesInserted / duration;
+      // eslint-disable-next-line no-console
       console.log('Throughput:', throughput, 'entries / second');
     }
   }
@@ -67,10 +70,12 @@ async function main() {
       mapSize: MAP_SIZE,
     });
     {
+      // eslint-disable-next-line no-console
       console.log('Generating entries for testing');
       const entries = [...Array(NUM_ENTRIES)].map(() => {
         return generateEntry();
       });
+      // eslint-disable-next-line no-console
       console.log('(manual batching) Writing entries for', MAX_TIME, 'ms');
       const start = Date.now();
       let numEntriesInserted = 0;
@@ -87,12 +92,14 @@ async function main() {
       }
       const duration = Date.now() - start;
       const throughput = numEntriesInserted / duration;
+      // eslint-disable-next-line no-console
       console.log('Safe Throughput:', throughput, 'entries / second');
     }
   }
 }
 
 main().catch((err) => {
+  // eslint-disable-next-line no-console
   console.error(err);
   process.exitCode = 1;
 });