+++ packages/core/atlaspack/static/runtime/browser-hmr/index.js
@@ -1,0 +1,1 @@
+module.exports = require('@atlaspack/runtime-browser-hmr');