+++ packages/core/eslint-plugin/test/utils.ts
@@ -1,7 +1,7 @@
 import * as mocha from 'mocha';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import tseslint from 'typescript-eslint';
 import {RuleTester} from '@typescript-eslint/rule-tester';
-import path from 'node:path';
 
 RuleTester.afterAll = mocha.after;