+++ packages/core/feature-flags/src/index.js
@@ -41,6 +41,7 @@ export const DEFAULT_FEATURE_FLAGS: FeatureFlags = {
   applyScopeHoistingImprovement: false,
   inlineConstOptimisationFix: false,
   hmrImprovements: false,
+  atlaspackV3CleanShutdown: false,
 };
 
 let featureFlagValues: FeatureFlags = {...DEFAULT_FEATURE_FLAGS};