+++ packages/core/integration-tests/test/javascript.js
@@ -30,6 +30,12 @@ import Logger from '@atlaspack/logger';
 import nullthrows from 'nullthrows';
 import {md} from '@atlaspack/diagnostic';
 
+// TS-MIGRATION: Replace with conditional exports
+let helperPath =
+  process.env.ATLASPACK_REGISTER_USE_SRC === 'true'
+    ? '@atlaspack/transformer-js/src/JSTransformer.js'
+    : '@atlaspack/transformer-js/lib/JSTransformer.js';
+
 describe('javascript', function () {
   beforeEach(async () => {
     await removeDistDirectory();
@@ -3968,6 +3974,7 @@ describe('javascript', function () {
         __dirname,
         'integration/undeclared-external/package.json',
       );
+
       await assert.rejects(
         () =>
           bundle(fixture, {
@@ -3981,9 +3988,7 @@ describe('javascript', function () {
           diagnostics: [
             {
               message: md`Failed to resolve '${'@swc/helpers/cjs/_class_call_check.cjs'}' from '${normalizePath(
-                require.resolve(
-                  '@atlaspack/transformer-js/src/JSTransformer.js',
-                ),
+                require.resolve(helperPath),
               )}'`,
               origin: '@atlaspack/core',
               codeFrames: [
@@ -4061,6 +4066,7 @@ describe('javascript', function () {
       );
       await overlayFS.mkdirp(path.dirname(pkg));
       await overlayFS.writeFile(pkg, pkgContents);
+
       await assert.rejects(
         () =>
           bundle(fixture, {
@@ -4075,9 +4081,7 @@ describe('javascript', function () {
           diagnostics: [
             {
               message: md`Failed to resolve '${'@swc/helpers/cjs/_class_call_check.cjs'}' from '${normalizePath(
-                require.resolve(
-                  '@atlaspack/transformer-js/src/JSTransformer.js',
-                ),
+                require.resolve(helperPath),
               )}'`,
               origin: '@atlaspack/core',
               codeFrames: [