+++ packages/core/package-manager/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @atlaspack/package-manager
 
+## 2.14.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atlaspack/fs@2.15.12
+  - @atlaspack/logger@2.14.12
+  - @atlaspack/utils@2.16.1
+  - @atlaspack/node-resolver-core@3.5.17
+  - @atlaspack/workers@2.14.17
+  - @atlaspack/types@2.15.7
+
+## 2.14.16
+
+### Patch Changes
+
+- Updated dependencies [[`30ee2cf`](https://github.com/atlassian-labs/atlaspack/commit/30ee2cfcd34cf2646ded0eda13fdb80a2a5de529)]:
+  - @atlaspack/utils@2.16.0
+  - @atlaspack/fs@2.15.11
+  - @atlaspack/node-resolver-core@3.5.16
+  - @atlaspack/workers@2.14.16
+  - @atlaspack/types@2.15.6
+
 ## 2.14.15
 
 ### Patch Changes